implement Robot_helper library
