+++ src/CMakeLists.txt
@@ -7,3 +7,4 @@ add_subdirectory(libiDynTree_helper)
 add_subdirectory(libTime_profiler)
 add_subdirectory(libStd_helper)
 add_subdirectory(libSimplified_model_controllers)
+add_subdirectory(libRobot_helper)
